size and sizeNoEnergy were int instead of float. Thats why there were no more particles anymore in some cases.

+++ source/glest_game/graphics/particle_type.h
@@ -62,8 +62,8 @@ protected:
 	Vec3f offset;
 	Vec4f color;
 	Vec4f colorNoEnergy;
-	int size;
-	int sizeNoEnergy;
+	float size;
+	float sizeNoEnergy;
 	float speed;
 	float gravity;
 	int emissionRate;
